move SpinUpAndShoot command
