+++ src/main/java/frc/robot/shooter/commands/SpinUpAndShoot.java
@@ -1,11 +1,10 @@
-package frc.robot.cmdGroup;
+package frc.robot.shooter.commands;
 
 import java.util.function.BooleanSupplier;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants;
 import frc.robot.shooter.ShooterSubsystem;
-import frc.robot.shooter.commands.*;
 
 public class SpinUpAndShoot extends SequentialCommandGroup {
   public SpinUpAndShoot(ShooterSubsystem shooter, BooleanSupplier slowShoot) {
